feat: 기물이 팀을 가지는 기능 구현
