+++ src/main/java/chess/domain/chessboard/state/Empty.java
@@ -1,10 +1,14 @@
 package chess.domain.chessboard.state;
 
-public final class Empty implements PieceState{
+public final class Empty implements PieceState {
 
     @Override
-    public boolean isEmpty(){
+    public boolean isEmpty() {
         return true;
     }
 
+    @Override
+    public Team getTeam() {
+        throw new UnsupportedOperationException();
+    }
 }
