--- src/main/java/chess/domain/chessboard/state/Pawn.java
@@ -1,8 +1,0 @@
-package chess.domain.chessboard.state;
-
-public final class Pawn extends Piece{
-
-    public Pawn(final Team team) {
-        super(team);
-    }
-}
