+++ src/main/java/chess/domain/chessboard/state/piece/Bishop.java
@@ -1,0 +1,10 @@
+package chess.domain.chessboard.state.piece;
+
+import chess.domain.chessboard.state.Team;
+
+public final class Bishop extends Piece {
+
+    public Bishop(final Team team) {
+        super(team);
+    }
+}
