+++ src/main/java/chess/domain/chessboard/state/piece/King.java
@@ -1,0 +1,10 @@
+package chess.domain.chessboard.state.piece;
+
+import chess.domain.chessboard.state.Team;
+
+public final class King extends Piece {
+
+    public King(final Team team) {
+        super(team);
+    }
+}
