+++ src/main/java/chess/domain/chessboard/state/piece/Knight.java
@@ -1,0 +1,10 @@
+package chess.domain.chessboard.state.piece;
+
+import chess.domain.chessboard.state.Team;
+
+public final class Knight extends Piece {
+
+    public Knight(final Team team) {
+        super(team);
+    }
+}
