+++ src/main/java/chess/domain/chessboard/state/piece/Pawn.java
@@ -1,0 +1,10 @@
+package chess.domain.chessboard.state.piece;
+
+import chess.domain.chessboard.state.Team;
+
+public final class Pawn extends Piece {
+
+    public Pawn(final Team team) {
+        super(team);
+    }
+}
