+++ src/main/java/chess/domain/chessboard/state/piece/Queen.java
@@ -1,0 +1,10 @@
+package chess.domain.chessboard.state.piece;
+
+import chess.domain.chessboard.state.Team;
+
+public final class Queen extends Piece {
+
+    public Queen(final Team team) {
+        super(team);
+    }
+}
