+++ src/main/java/chess/domain/chessboard/state/piece/Rook.java
@@ -1,0 +1,10 @@
+package chess.domain.chessboard.state.piece;
+
+import chess.domain.chessboard.state.Team;
+
+public final class Rook extends Piece {
+
+    public Rook(final Team team) {
+        super(team);
+    }
+}
